fix:main: Add SubscriptionFilterHelper
