+++ ewm-main/src/main/java/ru/practicum/ewm/subscription/model/filter/SubscriptionFilter.java
@@ -1,0 +1,15 @@
+package ru.practicum.ewm.subscription.model.filter;
+
+import lombok.Value;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Value
+public class SubscriptionFilter {
+    List<Long> categoryIds;
+    LocalDateTime rangeStart;
+    LocalDateTime rangeEnd;
+    Boolean paid;
+    Boolean onlyAvailable;
+}
