+++ ewm-main/src/main/java/ru/practicum/ewm/subscription/model/filter/SubscriptionFilterHelper.java
@@ -1,24 +1,16 @@
-package ru.practicum.ewm.subscription.model;
+package ru.practicum.ewm.subscription.model.filter;
 
 import com.querydsl.core.BooleanBuilder;
-import lombok.Value;
 import ru.practicum.ewm.event.model.QEvent;
 
-import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
-@Value
-public class SubscriptionFilter {
-    List<Long> categoryIds;
-    LocalDateTime rangeStart;
-    LocalDateTime rangeEnd;
-    Boolean paid;
-    Boolean onlyAvailable;
+public class SubscriptionFilterHelper {
 
-    public Optional<BooleanBuilder> makeBooleanBuilder(List<Long> favoriteIds, SubscriptionFilter filter) {
+    public static Optional<BooleanBuilder> makeBooleanBuilder(List<Long> favoriteIds, SubscriptionFilter filter) {
         java.util.function.Predicate<Object> isNullOrEmpty = obj ->
                 Objects.isNull(obj) || (obj instanceof Collection && ((Collection<?>) obj).isEmpty());
         QEvent qEvent = QEvent.event;
